requirement.txt
streamlit
crewai
crew[tools]
langchain_core
python-dotenv
langchain_community
invoke
tools